+++ src/viam/media/__init__.py
@@ -1,0 +1,9 @@
+from .media import MediaReader, MediaSource, MediaStream, MediaStreamWithIterator, MediaType
+
+__all__ = [
+    "MediaReader",
+    "MediaSource",
+    "MediaStream",
+    "MediaStreamWithIterator",
+    "MediaType",
+]
